+++ CHANGELOG.md
@@ -8,6 +8,7 @@ develop
 ### Improvements
 
 - Tweak preferences layout.
+- Add dot mark to installed syntax style list in preferences to represent the state if style is customized.
 
 
 ### Fixes